migrate angle utils from float to double

+++ 01_Libraries/ghost_util/include/ghost_util/angle_util.hpp
@@ -4,15 +4,15 @@
 
 namespace ghost_util {
 
-float WrapAngle360(float angle);
-float WrapAngle180(float angle);
-float FlipAngle180(float angle);
+double WrapAngle360(double angle);
+double WrapAngle180(double angle);
+double FlipAngle180(double angle);
 
-float WrapAngle2PI(float angle);
-float WrapAnglePI(float angle);
-float FlipAnglePI(float angle);
+double WrapAngle2PI(double angle);
+double WrapAnglePI(double angle);
+double FlipAnglePI(double angle);
 
-float SmallestAngleDistDeg(float a2, float a1);
-float SmallestAngleDistRad(float a2, float a1);
+double SmallestAngleDistDeg(double a2, double a1);
+double SmallestAngleDistRad(double a2, double a1);
 
 } // namespace ghost_util
